+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - GAP: return command disallowed if disconnect already requested
 - GAP: improve handling of incorrectly resolved addresses in HCI_SUBEVENT_LE_CONNECTION_COMPLETE
+- GAP: only store link key if at least one side requests bonding during the IO Capabilities exchange.
 - GOEP Client: remove goep_client_create_connection. Use goep_client_connect instead.
 - HID Parser: cleanup of function names and signatures
 - HIDS Client: use error code instead of att status in conencted event